Default profile image is uploaded upon signup to Firebase storage; annotation images are now unique; updated UserViewController and LogIn view controller to accomodate for these changes

+++ Pitch/Views/Explore Views/NumberOfPeopleCell.h
@@ -18,8 +18,8 @@ NS_ASSUME_NONNULL_BEGIN
 @property (strong, nonatomic) MARKRangeSlider *rangeSlider;
 
 - (void) resetNumberOfPeople;
-- (int) getMinNumPeople;
-- (int) getMaxNumPeople;
+- (long) getMinNumPeople;
+- (long) getMaxNumPeople;
 
 @end
